+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/panel/FormBuilderTest.java
@@ -276,7 +276,7 @@ public final class FormBuilderTest {
 
 	private static JTextField field(String label) {
 		return TextFieldBuilder.builder()
-						.valueClass(String.class)
+						.type(String.class)
 						.label(label)
 						.build();
 	}